Tentative to solve sps_lut problems in 64 Bit machines. Make sure Numeric.zeros gets integers in PyMca/PyMcaPostBatch.py
git-svn-id: https://svn.code.sf.net/p/pymca/code@104 a8702ac3-0063-49a2-b157-4a214673c9c1

+++ PyMca/sps/Include/sps_lut.h
@@ -41,9 +41,9 @@
 typedef struct XServer_Info {
   int byte_order;
   int pixel_size;
-  unsigned long red_mask;
-  unsigned long green_mask;
-  unsigned long blue_mask;
+  unsigned int red_mask;
+  unsigned int green_mask;
+  unsigned int blue_mask;
 } XServer_Info;
 
 /* GILLES DOC is as usual absent ... but at least these are the different
